+++ endavaRestApi/Repositories/IOrderRepository.cs
@@ -7,6 +7,6 @@ namespace endavaRestApi.Repositories
         Task<bool> IsUserActive(int userId);
         Task<bool> AreProductsAvailable(Dictionary<int, int> productQuantities);
         Task<Order> CreateOrder(int userId, Dictionary<int, int> productQuantities);
-        Task<Payment> CreatePayment(int orderId, decimal amount);
+        Task<Payment> CreatePayment(int orderId, int amount);
     }
 }
